// src/api/models/ClientesNaturalRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientesNaturalRequest = {
    tipoDocumento: string;
    numeroIdentificacion: string;
    nombres: string;
    apellidos: string;
    fechaNacimiento: string;
    genero: string;
};